Module Low_level.Process

type t

A process

val pid : t -> int

pid t returns the process id of t.

val to_parent_pipe : ?readable_in_child:bool -> ?writable_in_child:bool -> ?overlapped:bool -> fd:int -> parent_pipe:Pipe.t -> unit -> Luv.Process.redirection

Wraps Luv.Process.to_parent_pipe

val await_exit : t -> int * int64

await_exit t waits for the process t to finish.

It returns (exit_status, term_signal), see Luv.Process.spawn for more details on these values.

val has_exited : t -> bool

has_exited t checks if the process t has exited or not.

val send_signal : t -> int -> unit

A wrapper for Luv.Process.kill.

val spawn : ?cwd:string -> ?env:(string * string) list -> ?uid:int -> ?gid:int -> ?redirect:Luv.Process.redirection list -> sw:Eio.Std.Switch.t -> string -> string list -> t

Wraps Luv.Process.spawn.

The process will be stopped when the switch is released if it has not already exited.