Module Low_level.Stream

type 'a t = [ `Stream of 'a ] Handle.t

val read_into : [< `Pipe | `TCP | `TTY ] t -> Luv.Buffer.t -> int

read_into handle buf reads some bytes from handle into buf returning the number of bytes read.

raises End_of_file
if there is no more data to read

val write : [ `Stream of [< `Pipe | `TCP | `TTY ] ] Handle.t -> Luv.Buffer.t list -> unit

write handle bufs writes the contents of bufs to handle.